requirements.txt
pre-commit
torch
torchvision
pytorch-lightning
pytorch-lightning-bolts
wandb
hydra-core
